first pass of migrating slots to DocumentWindow class out of DocumentController
