+++ cadnano/views/documentwindow.py
@@ -3,35 +3,58 @@ from collections import namedtuple
 
 from PyQt5.QtCore import (
     Qt,
-    QSettings,
+    QCoreApplication,
+    QDir,
+    QEvent,
+    QFileInfo,
     QPoint,
+    QRect,
+    QSettings,
     QSize,
-    QEvent
 )
+_translate = QCoreApplication.translate
 from PyQt5.QtGui import (
     QMoveEvent,
     QResizeEvent
 )
 from PyQt5.QtWidgets import (
     QAction,
+    QActionGroup,
     QApplication,
-    QWidget,
+    QDialog,
+    QFileDialog,
     QGraphicsItem,
+    QGraphicsScene,
+    QMessageBox,
     QMainWindow,
-    QGraphicsScene
+    QStyleOptionGraphicsItem,
+    QWidget
 )
+from PyQt5.QtGui import (
+    QKeySequence,
+    QPainter,
+    QCloseEvent
+)
+from PyQt5.QtSvg import QSvgGenerator
 
-
-from cadnano import app
+from cadnano import (
+    app,
+    setReopen,
+    util
+)
+from cadnano.views import styles
 from cadnano.fileio.v3encode import reEmitPart
 from cadnano.proxies.cnproxy import UndoStack
 from cadnano.gui.mainwindow import ui_mainwindow
+from cadnano.gui.dialogs.ui_about import Ui_About
 from cadnano.proxies.cnenum import (
+    EnumType,
+    GridEnum,
     OrthoViewEnum,
     ViewReceiveEnum,
     ViewSendEnum,
-    EnumType
 )
+from cadnano.controllers.documentcontroller import DocumentController
 from cadnano.views.gridview.gridrootitem import GridRootItem
 from cadnano.views.gridview.tools.gridtoolmanager import GridToolManager
 from cadnano.views.pathview.colorpanel import ColorPanel
@@ -46,6 +69,15 @@ from cadnano.cntypes import (
     GraphicsViewT
 )
 
+IS_TESTING = True
+
+DEFAULT_VHELIX_FILTER = True
+ONLY_ONE = True  # bool: Retricts Document to creating only one Part if True.
+SAVE_DIALOG_OPTIONS = {'SAVE': 0,
+                       'CANCEL': 1,
+                       'DISCARD': 2
+                       }
+
 # from PyQt5.QtOpenGL import QGLWidget
 # # check out https://github.com/baoboa/pyqt5/tree/master/examples/opengl
 # # for an example of the QOpenGlWidget added in Qt 5.4
@@ -58,21 +90,45 @@ class DocumentWindow(QMainWindow, ui_mainwindow.Ui_MainWindow):
     Attributes:
         controller: DocumentController
     """
+    filter_list = ["strand", "endpoint", "xover", "virtual_helix"]
 
-    def __init__(self, doc_ctrlr: DocCtrlT, parent=None):
+    def __init__(self, document: DocT, parent=None):
         super(DocumentWindow, self).__init__(parent)
+        self._document: DocT = document
+        self.controller: DocumentController = DocumentController(self, document)
 
-        self.controller = doc_ctrlr
-        doc = doc_ctrlr.document()
         self.setupUi(self)
-        self.settings = QSettings("cadnano.org", "cadnano2.5")
+
+        self.fileopendialog = None
+        self.filesavedialog = None
+        self._file_open_path = None  # will be set in _readSettings
+        self._has_no_associated_file: bool = True
+
+        # self.settings = QSettings("cadnano.org", "cadnano2.5")
+        self.settings = QSettings(  "69bfae41ee5e33c689fded70c89cc64c",
+                                    "69bfae41ee5e33c689fded70c89cc64c")
+        self._readSettings()
+
+        self._hintable_tools: List = []     # filters that display alt icon when disabled
+        self._hintable_filters: List = []   # filters that display alt icon when disabled
+        self._hintable_tool_action_map: dict = {}  # what buttons to hint for each filter
+        self._hintable_filter_action_map: dict = {}  # what buttons to hint for each filter
+        self._tool_hints_visible: bool = False
+        self._filter_hints_visible: bool = False
+
+        self.slice_view_showing: bool = False
+
         # Appearance pref
         if not app().prefs.show_icon_labels:
             self.main_toolbar.setToolButtonStyle(Qt.ToolButtonIconOnly)
 
         # Outliner & PropertyEditor setup
-        self.outliner_widget.configure(window=self, document=doc)
-        self.property_widget.configure(window=self, document=doc)
+        outliner_widget = self.outliner_widget
+        outliner_widget.configure(window=self, document=doc)
+        property_widget = self.property_widget
+        property_widget.configure(window=self, document=doc)
+        outliner_widget.itemSelectionChanged.connect(property_widget.outlinerItemSelectionChanged)
+
         self.property_buttonbox.setVisible(False)
 
         self.tool_managers = None  # initialize
@@ -99,10 +155,64 @@ class DocumentWindow(QMainWindow, ui_mainwindow.Ui_MainWindow):
         self._finishInit()
 
         doc.setViewNames(['slice', 'path', 'inspector'])
+
+        app().documentWindowWasCreatedSignal.emit(self._document, self)
+
+        # Set Default Filter
+        if DEFAULT_VHELIX_FILTER:
+            self.actionFilterVirtualHelixSlot()
+        else:
+            self.actionFilterEndpointSlot()
+            self.actionFilterXoverSlot()
+
+                # setup tool exclusivity
+        self.actiongroup = ag = QActionGroup(self)
+        action_group_list = ['action_global_select',
+                             'action_global_create',
+                             'action_path_break',
+                             'action_path_paint',
+                             'action_path_insertion',
+                             'action_path_skip',
+                             'action_path_add_seq',
+                             'action_path_mods']
+        for action_name in action_group_list:
+            ag.addAction(getattr(self, action_name))
+
+        if IS_TESTING:
+            # ADDED SPECIAL NC for testing with keyboard shortcut:
+            action_special = QAction(self)
+            action_special.setObjectName("action_special")
+            action_special.setText(_translate("MainWindow", "Special"))
+            action_special.setShortcut(_translate("MainWindow", "Ctrl+J"))
+            self.action_special = action_special
+            self.menu_edit.addAction(action_special)
+            action_special.triggered.connect(self.actionSpecialSlot)
+
+        # set up tool & filter hinting
+        self._hintable_tools = [self.action_global_create,
+                                self.action_global_select]
+        self._hintable_filters = [self.action_filter_helix,
+                                  self.action_filter_strand,
+                                  self.action_filter_endpoint,
+                                  self.action_filter_xover]
+        self._hintable_tool_action_map = {'create': [self.action_global_create],
+                                          'select': [self.action_global_select]}
+
+        self._hintable_filter_action_map = {'virtual_helix': [self.action_filter_helix],
+                                            'strand': [self.action_filter_strand],
+                                            'endpoint': [self.action_filter_endpoint],
+                                            'xover': [self.action_filter_xover]}
+
+        self.action_global_select.trigger()
+        # self.inspector_dock_widget.hide()
+
+        self.exit_when_done = False
+        self.closeEvent = self.windowCloseEventHandler
+        self.show()
     # end def
 
     def document(self) -> DocT:
-        return self.controller.document()
+        return self._document
 
     def destroyWin(self):
         '''Save window state and destroy the tool managers.  Also destroy
@@ -400,6 +510,897 @@ class DocumentWindow(QMainWindow, ui_mainwindow.Ui_MainWindow):
             if view.underMouse():
                 self.destroyView(name)
     # end def
+
+    ##### SLOTS
+    def actionSpecialSlot(self):
+        self.win.doMouseViewDestroy()
+    # end def
+
+    def actionSelectForkSlot(self):
+        win = self.win
+        win.action_path_select.trigger()
+        win.action_vhelix_select.trigger()
+    # end def
+
+    def actionCreateForkSlot(self):
+        self._document.clearAllSelected()
+        win = self.win
+        win.action_vhelix_create.trigger()
+        win.action_path_create.trigger()
+    # end def
+
+    def actionVhelixSnapSlot(self, state: bool):
+        for item in self.win.sliceroot.instance_items.values():
+            item.griditem.allow_snap = state
+    # end def
+
+    def undoStackCleanChangedSlot(self):
+        '''The title changes to include [*] on modification.
+        Use this when clearing out undostack to set the modified status of
+        the document
+        '''
+        self.win.setWindowModified(not self.undoStack().isClean())
+        self.win.setWindowTitle(self.documentTitle())
+    # end def
+
+    def actionAboutSlot(self):
+        '''Displays the about cadnano dialog.'''
+        dialog = QDialog()
+        dialog_about = Ui_About()  # reusing this dialog, should rename
+        dialog.setStyleSheet(
+            "QDialog { background-image: url(ui/dialogs/images/cadnano2-about.png); background-repeat: none; }")
+        dialog_about.setupUi(dialog)
+        dialog.exec_()
+
+    def showFilterHints(self, show_hints: bool, filter_name: str = None):
+        '''Changes the appearance of filter buttons in the toolbar to help user
+        realize they may not have the correct filters selected to perform a
+        desired action. Meant to avoid the case where user clicks and nothing
+        happens.
+
+        Buttons (:class:`QAction`s) can automatically display different icons
+        depending on state. We use "enabled" state, temporarily disabling the
+        :class:`QAction` to toggle its appearance.
+
+        View items should call this method once from a :meth:`mousePressEvent`
+        with ``show_hints=True``, and again on :meth:`mouseReleaseEvent` with
+        ``show_hints=False``.
+
+        Args:
+            show_hints: ``True to show hints, ``False`` to hide hints.
+            filter_name: What filter to hint from ``self._hintable_filter_action_map``
+        '''
+        if filter_name:
+            try:
+                for f in self._hintable_filter_action_map[filter_name]:
+                    if not f.isChecked():  # no need to hint when already checked
+                        f.setEnabled(False)  # disable to show hint icon
+                        self._filter_hints_visible = True
+            except KeyError:
+                pass
+        elif self._filter_hints_visible:
+            for f in self._hintable_filters:
+                if f.isEnabled() is False:
+                    f.setEnabled(True)  # enable to hide hint icon
+            self._filter_hints_visible = False
+
+    def showToolHints(self, show_hints: bool, tool_name: str = None):
+        if tool_name:
+            try:
+                for t in self._hintable_tool_action_map[tool_name]:
+                    if not t.isChecked():  # no need to hint when already checked
+                        t.setEnabled(False)  # disable to show hint icon
+                        self._tool_hints_visible = True
+            except KeyError:
+                pass
+        elif self._tool_hints_visible:
+            for t in self._hintable_tools:
+                if t.isEnabled() is False:
+                    t.setEnabled(True)  # enable to hide hint icon
+            self._tool_hints_visible = False
+    # end def
+
+    def actionFilterVirtualHelixSlot(self):
+        '''Disables all other selection filters when active.'''
+        win = self.win
+        fH = win.action_filter_helix
+        fE = win.action_filter_endpoint
+        fS = win.action_filter_strand
+        fX = win.action_filter_xover
+        fH.setChecked(True)
+        if fE.isChecked():
+            fE.setChecked(False)
+        if fS.isChecked():
+            fS.setChecked(False)
+        if fX.isChecked():
+            fX.setChecked(False)
+        types = ["virtual_helix"]
+        self._document.setFilterSet(types)
+    # end def
+
+    def actionFilterEndpointSlot(self):
+        '''Disables handle filters when activated.
+        Remains checked if no other item-type filter is active.
+        '''
+        win = self.win
+        fH = win.action_filter_helix
+        fE = win.action_filter_endpoint
+        fS = win.action_filter_strand
+        fX = win.action_filter_xover
+        if fH.isChecked():
+            fH.setChecked(False)
+        if not fS.isChecked() and not fX.isChecked():
+            fE.setChecked(True)
+            fS.setChecked(True)
+        self._strandFilterUpdate()
+    # end def
+
+    def actionFilterStrandSlot(self):
+        '''Disables handle filters when activated.
+        Remains checked if no other item-type filter is active.
+        '''
+        win = self.win
+        fH = win.action_filter_helix
+        fE = win.action_filter_endpoint
+        fS = win.action_filter_strand
+        fX = win.action_filter_xover
+        if fH.isChecked():
+            fH.setChecked(False)
+        if not fE.isChecked() and not fX.isChecked():
+            fS.setChecked(True)
+        self._strandFilterUpdate()
+    # end def
+
+    def actionFilterXoverSlot(self):
+        '''Disables handle filters when activated.
+        Remains checked if no other item-type filter is active.
+        '''
+        win = self.win
+        fH = win.action_filter_helix
+        fE = win.action_filter_endpoint
+        fS = win.action_filter_strand
+        fX = win.action_filter_xover
+        if fH.isChecked():
+            fH.setChecked(False)
+        if not fE.isChecked() and not fS.isChecked():
+            fX.setChecked(True)
+        self._strandFilterUpdate()
+    # end def
+
+    def actionFilterScafSlot(self):
+        '''Remains checked if no other strand-type filter is active.'''
+        f_scaf = self.win.action_filter_scaf
+        f_stap = self.win.action_filter_stap
+        if not f_scaf.isChecked() and not f_stap.isChecked():
+            f_scaf.setChecked(True)
+        self._strandFilterUpdate()
+
+    def actionFilterStapSlot(self):
+        '''Remains checked if no other strand-type filter is active.'''
+        f_scaf = self.win.action_filter_scaf
+        f_stap = self.win.action_filter_stap
+        if not f_scaf.isChecked() and not f_stap.isChecked():
+            f_stap.setChecked(True)
+        self._strandFilterUpdate()
+    # end def
+
+    def actionCopySlot(self):
+        win = self.win
+        select_tool = win.getMouseViewTool('select')
+        if select_tool is not None and hasattr(select_tool, 'copySelection'):
+            print("select_tool is rolling")
+            select_tool.copySelection()
+    # end def
+
+    def actionPasteSlot(self):
+        win = self.win
+        select_tool = win.getMouseViewTool('select')
+        if select_tool is not None and hasattr(select_tool, 'pasteClipboard'):
+            select_tool.pasteClipboard()
+    # end def
+
+    def actionFilterFwdSlot(self):
+        '''Remains checked if no other strand-type filter is active.'''
+        f_fwd = self.win.action_filter_fwd
+        f_rev = self.win.action_filter_rev
+        if not f_fwd.isChecked() and not f_rev.isChecked():
+            f_fwd.setChecked(True)
+        self._strandFilterUpdate()
+
+    def actionFilterRevSlot(self):
+        '''Remains checked if no other strand-type filter is active.'''
+        f_fwd = self.win.action_filter_fwd
+        f_rev = self.win.action_filter_rev
+        if not f_fwd.isChecked() and not f_rev.isChecked():
+            f_rev.setChecked(True)
+        self._strandFilterUpdate()
+    # end def
+
+    def _strandFilterUpdate(self):
+        win = self.win
+        if win.action_filter_helix.isChecked():
+            self._document.setFilterSet(["virtual_helix"])
+            return
+
+        filter_list = []
+        add_oligo = False
+        if win.action_filter_endpoint.isChecked():
+            filter_list.append("endpoint")
+            add_oligo = True
+        if win.action_filter_strand.isChecked():
+            filter_list.append("strand")
+            add_oligo = True
+        if win.action_filter_xover.isChecked():
+            filter_list.append("xover")
+            add_oligo = True
+        if win.action_filter_fwd.isChecked():
+            filter_list.append("forward")
+            add_oligo = True
+        if win.action_filter_rev.isChecked():
+            filter_list.append("reverse")
+            add_oligo = True
+        if win.action_filter_scaf.isChecked():
+            filter_list.append("scaffold")
+            add_oligo = True
+        if win.action_filter_stap.isChecked():
+            filter_list.append("staple")
+            add_oligo = True
+        if add_oligo:
+            filter_list.append("oligo")
+        self._document.setFilterSet(filter_list)
+    # end def
+
+    def actionNewSlot(self):
+        '''1. If document is has no parts, do nothing.
+        2. If document is dirty, call promptSaveDialog and continue if it succeeds.
+        3. Create a new document and swap it into the existing ctrlr/window.
+        '''
+        # clear/reset the view!
+
+        if len(self._document.children()) == 0:
+            return  # no parts
+        elif self.promptSaveDialog() is SAVE_DIALOG_OPTIONS['CANCEL']:
+            return  # user canceled in maybe save
+        else:  # user did not cancel
+            if self.filesavedialog is not None:
+                self.filesavedialog.finished.connect(self.newClickedCallback)
+            else:  # user did not save
+                self.newClickedCallback()  # finalize new
+
+    def actionOpenSlot(self):
+        '''1. If document is untouched, proceed to open dialog.
+        2. If document is dirty, call maybesave and continue if it succeeds.
+        3. Downstream, the file is selected in openAfterMaybeSave, and the selected
+           file is actually opened in openAfterMaybeSaveCallback.
+        '''
+        if self.promptSaveDialog() is SAVE_DIALOG_OPTIONS['CANCEL']:
+            return  # user canceled in maybe save
+        else:  # user did not cancel
+            if hasattr(self, "filesavedialog"):  # user did save
+                if self.filesavedialog is not None:
+                    self.filesavedialog.finished.connect(self.openAfterMaybeSave)
+                else:
+                    self.openAfterMaybeSave()  # windows
+            else:  # user did not save
+                self.openAfterMaybeSave()  # finalize new
+
+    def actionCloseSlot(self):
+        '''Called when DocumentWindow is closed.
+
+        This method does not do anything as its functionality is implemented by
+        windowCloseEventHandler.  windowCloseEventHandler captures all close
+        events (including quit signals) while this method captures only close
+        window events (i.e. a strict subset of the events that the
+        windowCloseEventHandler captures.
+        '''
+        pass
+
+    def actionSaveSlot(self):
+        '''SaveAs if necessary, otherwise overwrite existing file.'''
+        if self._has_no_associated_file:
+            self.saveFileDialog()
+            return
+        self.writeDocumentToFile()
+
+    def actionSaveAsSlot(self):
+        '''Open a save file dialog so user can choose a name.'''
+        self.saveFileDialog()
+
+    def actionSVGSlot(self):
+        ''''''
+        fname = os.path.basename(str(self.fileName()))
+        if fname is None:
+            directory = "."
+        else:
+            directory = QFileInfo(fname).path()
+
+        fdialog = QFileDialog(self.win,
+                              "%s - Save As" % QApplication.applicationName(),
+                              directory,
+                              "%s (*.svg)" % QApplication.applicationName())
+        fdialog.setAcceptMode(QFileDialog.AcceptSave)
+        fdialog.setWindowFlags(Qt.Sheet)
+        fdialog.setWindowModality(Qt.WindowModal)
+        self.svgsavedialog = fdialog
+        self.svgsavedialog.filesSelected.connect(self.saveSVGDialogCallback)
+        fdialog.open()
+
+    class DummyChild(QGraphicsItem):
+        def boundingRect(self):
+            return QRect(200, 200)  # self.parentObject().boundingRect()
+
+        def paint(self, painter, option, widget=None):
+            pass
+
+    def saveSVGDialogCallback(self, selected: Union[str, list, tuple]):
+        if isinstance(selected, (list, tuple)):
+            fname = selected[0]
+        else:
+            fname = selected
+        if fname is None or os.path.isdir(fname):
+            return False
+        if not fname.lower().endswith(".svg"):
+            fname += ".svg"
+        if self.svgsavedialog is not None:
+            self.svgsavedialog.filesSelected.disconnect(self.saveSVGDialogCallback)
+            del self.svgsavedialog  # prevents hang
+            self.svgsavedialog = None
+
+        generator = QSvgGenerator()
+        generator.setFileName(fname)
+        generator.setSize(QSize(200, 330))
+        generator.setViewBox(QRect(0, 0, 2000, 2000))
+        painter = QPainter()
+
+        # Render through scene
+        # painter.begin(generator)
+        # self.win.pathscene.render(painter)
+        # painter.end()
+
+        # Render item-by-item
+        painter = QPainter()
+        style_option = QStyleOptionGraphicsItem()
+        q = [self.win.pathroot]
+        painter.begin(generator)
+        while q:
+            graphics_item = q.pop()
+            transform = graphics_item.itemTransform(self.win.sliceroot)[0]
+            painter.setTransform(transform)
+            if graphics_item.isVisible():
+                graphics_item.paint(painter, style_option, None)
+                q.extend(graphics_item.childItems())
+        painter.end()
+
+    def actionExportSequencesSlot(self):
+        '''Triggered by clicking Export Staples button. Opens a file dialog to
+        determine where the staples should be saved. The callback is
+        exportStaplesCallback which collects the staple sequences and exports
+        the file.
+        '''
+        # Validate that no staple oligos are circular.
+        part = self._document.activePart()
+        if part is None:
+            return
+        circ_olgs = part.getCircularOligos()
+        if circ_olgs:
+            from cadnano.gui.dialogs.ui_warning import Ui_Warning
+            dialog = QDialog()
+            dialogWarning = Ui_Warning()  # reusing this dialog, should rename
+            dialog.setStyleSheet(
+                "QDialog { background-image: url(ui/dialogs/images/cadnano2-about.png); background-repeat: none; }")
+            dialogWarning.setupUi(dialog)
+
+            locs = ", ".join([o.locString() for o in circ_olgs])
+            msg = "Part contains staple loop(s) at %s.\n\nUse the break tool to introduce 5' & 3' ends before exporting. Loops have been colored red; use undo to revert." % locs  # noqa
+            dialogWarning.title.setText("Staple validation failed")
+            dialogWarning.message.setText(msg)
+            for o in circ_olgs:
+                o.applyColor(styles.stapColors[0])
+            dialog.exec_()
+            return
+
+        # Proceed with staple export.
+        fname = self.fileName()
+        if fname is None:
+            directory = "."
+        else:
+            directory = QFileInfo(fname).path()
+        if util.isWindows():  # required for native looking file window
+            fname = QFileDialog.getSaveFileName(self.win,
+                                                "%s - Export As" % QApplication.applicationName(),
+                                                directory,
+                                                "(*.txt)")
+            self.saveStaplesDialog = None
+            self.exportStaplesCallback(fname)
+        else:  # access through non-blocking callback
+            fdialog = QFileDialog(self.win,
+                                  "%s - Export As" % QApplication.applicationName(),
+                                  directory,
+                                  "(*.txt)")
+            fdialog.setAcceptMode(QFileDialog.AcceptSave)
+            fdialog.setWindowFlags(Qt.Sheet)
+            fdialog.setWindowModality(Qt.WindowModal)
+            self.saveStaplesDialog = fdialog
+            self.saveStaplesDialog.filesSelected.connect(self.exportStaplesCallback)
+            fdialog.open()
+    # end def
+
+    def actionPathAddSeqSlot(self):
+        pass
+        # print("triggered seqslot")
+        # ap = self._document.activePart()
+        # if ap is not None:
+        #     self._document.activePart().setAbstractSequences(emit_signals=True)
+    # end def
+
+    def actionPrefsSlot(self):
+        app().prefsClicked()
+    # end def
+
+    def actionModifySlot(self):
+        '''Notifies that part root items that parts should respond to modifier
+        selection signals.
+        '''
+
+    def actionCreateNucleicAcidPartHoneycomb(self):
+        # TODO[NF]:  Docstring
+        self.actionCreateNucleicAcidPart(grid_type=GridEnum.HONEYCOMB)
+
+    def actionCreateNucleicAcidPartSquare(self):
+        # TODO[NF]:  Docstring
+        self.actionCreateNucleicAcidPart(grid_type=GridEnum.SQUARE)
+
+    def actionCreateNucleicAcidPart(self, grid_type: EnumType) -> NucleicAcidPartT:
+        # TODO[NF]:  Docstring
+        if ONLY_ONE:
+            if len(self._document.children()) is not 0:
+                if self.promptSaveDialog() is SAVE_DIALOG_OPTIONS['CANCEL']:
+                    return
+            self.newDocument()
+        doc = self._document
+        part = doc.createNucleicAcidPart(use_undostack=True, grid_type=grid_type)
+        active_part = doc.activePart()
+        if active_part is not None:
+            active_part.setActive(False)
+            doc.deactivateActivePart()
+        part.setActive(True)
+        doc.setActivePart(part)
+        return part
+    # end def
+
+    def actionToggleSliceViewSlot(self):
+        '''Handle the action_slice button being clicked.
+
+        If either the slice or grid view are visible hide them.  Else, revert
+        to showing whichever view(s) are showing.
+        '''
+
+        win = self.win
+        if win.action_slice.isChecked():
+        # if not (win.slice_dock_widget.isVisible() or win.grid_dock_widget.isVisible()):
+            win.action_slice.setChecked(True)
+            if self.slice_view_showing:
+                self.toggleSliceView(True)
+                self.toggleGridView(False)
+            else:
+                self.toggleGridView(True)
+                self.toggleSliceView(False)
+        else:
+            win.action_slice.setChecked(False)
+            self.toggleSliceView(False)
+            self.toggleGridView(False)
+
+    # end def
+
+    def actionTogglePathViewSlot(self):
+        dock_window = self.win.path_dock_widget
+        if dock_window.isVisible():
+            dock_window.hide()
+        else:
+            dock_window.show()
+    # end def
+
+    def actionToggleInspectorViewSlot(self):
+        dock_window = self.win.inspector_dock_widget
+        if dock_window.isVisible():
+            dock_window.hide()
+        else:
+            dock_window.show()
+    # end def
+
+    def toggleNewPartButtons(self, is_enabled: bool):
+        '''Toggle the AddPart buttons when the active part changes.
+
+        Args:
+            is_enabled: Whether the slice view should be hidden or shown
+        '''
+        self.win.action_new_dnapart_honeycomb.setEnabled(is_enabled)
+        self.win.action_new_dnapart_square.setEnabled(is_enabled)
+    # end def
+
+    def setSliceOrGridViewVisible(self, view_type: EnumType):
+        '''
+        Args:
+            view_type: type of view enum
+
+        Raises:
+            ValueError for unknown ``view_type``.
+        '''
+        if view_type == OrthoViewEnum.SLICE:
+            self.slice_view_showing = True
+        elif view_type == OrthoViewEnum.GRID:
+            self.slice_view_showing = False
+        else:
+            raise ValueError('Invalid orthoview value: %s' % value)
+        self.actionToggleSliceViewSlot()
+    # end def
+
+    def toggleSliceView(self, show: bool):
+        '''Hide or show the slice view based on the given parameter `show`.
+
+        Since calling this method where show=True will cause the SliceView to
+        show, ensure that the action_slice button is checked if applicable.
+
+        Args:
+            show: Whether the slice view should be hidden or shown
+
+        '''
+        assert isinstance(show, bool)
+
+        slice_view_widget = self.win.slice_dock_widget
+        path_view_widget = self.win.path_dock_widget
+        views = self.win.views
+        sgv = views[ViewSendEnum.SLICE]
+        if show:
+            self.win.splitDockWidget(slice_view_widget, path_view_widget, Qt.Horizontal)
+            # self.win.splitDockWidget(slice_view_widget, path_view_widget, Qt.Vertical)
+            slice_view_widget.show()
+            sgv.zoomToFit() # NOTE ZTF for now rather than copying the scale factor
+        else:
+            slice_view_widget.hide()
+    # end def
+
+    def toggleGridView(self, show: bool):
+        '''Hide or show the grid view based on the given parameter `show`
+
+        Since calling this method where show=True will cause the SliceView to
+        show, ensure that the action_slice button is checked if applicable.
+
+        Args:
+            show: Whether the grid view should be hidden or shown
+        '''
+        assert isinstance(show, bool)
+        grid_view_widget = self.win.grid_dock_widget
+        path_view_widget = self.win.path_dock_widget
+        views = self.win.views
+        ggv = views[ViewSendEnum.GRID]
+        if show:
+            self.win.splitDockWidget(grid_view_widget, path_view_widget, Qt.Horizontal)
+            # self.win.splitDockWidget(grid_view_widget, path_view_widget, Qt.Vertical)
+            grid_view_widget.show()
+            ggv.zoomToFit() # NOTE ZTF for now rather than copying the scale factor
+        else:
+            grid_view_widget.hide()
+    # end def
+
+    ### ACCESSORS ###
+    def document(self) -> DocT:
+        return self._document
+    # end def
+
+    def window(self) -> WindowT:
+        return self.win
+    # end def
+
+    def setDocument(self, doc: DocT):
+        '''Sets the controller's document, and informs the document that
+        this is its controller.
+        '''
+        self._document = doc
+        doc.setController(self)
+    # end def
+
+    def undoStack(self) -> UndoStack:
+        return self._document.undoStack()
+    # end def
+
+    ### PRIVATE SUPPORT METHODS ###
+    def newDocument(self, fname: str = None):
+        '''Creates a new Document, reusing the DocumentController.
+        Tells all of the views to reset and removes all items from
+        them
+        '''
+        if fname is not None and self.fileName() == fname:
+            setReopen(True)
+        self._document.makeNew()
+        self._has_no_associated_file = fname is None
+        self.win.setWindowTitle(self.documentTitle() + '[*]')
+    # end def
+
+    def saveFileDialog(self):
+        fname = self.fileName()
+        if fname is None:
+            directory = "."
+        else:
+            directory = QFileInfo(fname).path()
+        if util.isWindows():  # required for native looking file window
+            fname = QFileDialog.getSaveFileName(self.win,
+                                                "%s - Save As" % QApplication.applicationName(),
+                                                directory,
+                                                "%s (*.json)" % QApplication.applicationName())
+            if isinstance(fname, (list, tuple)):
+                fname = fname[0]
+            self.writeDocumentToFile(fname)
+        else:  # access through non-blocking callback
+            fdialog = QFileDialog(self.win,
+                                  "%s - Save As" % QApplication.applicationName(),
+                                  directory,
+                                  "%s (*.json)" % QApplication.applicationName())
+            fdialog.setAcceptMode(QFileDialog.AcceptSave)
+            fdialog.setWindowFlags(Qt.Sheet)
+            fdialog.setWindowModality(Qt.WindowModal)
+            self.filesavedialog = fdialog
+            self.filesavedialog.filesSelected.connect(self.saveFileDialogCallback)
+            fdialog.open()
+    # end def
+
+    def _readSettings(self):
+        self.settings.beginGroup("FileSystem")
+        self._file_open_path = self.settings.value("openpath", QDir().homePath())
+        self.settings.endGroup()
+
+    def _writeFileOpenPath(self, path):
+        ''''''
+        self._file_open_path = path
+        self.settings.beginGroup("FileSystem")
+        self.settings.setValue("openpath", path)
+        self.settings.endGroup()
+
+    ### SLOT CALLBACKS ###
+    def actionNewSlotCallback(self):
+        '''Gets called on completion of filesavedialog after newClicked's
+        promptSaveDialog. Removes the dialog if necessary, but it was probably
+        already removed by saveFileDialogCallback.
+        '''
+        if self.filesavedialog is not None:
+            self.filesavedialog.finished.disconnect(self.actionNewSlotCallback)
+            del self.filesavedialog  # prevents hang (?)
+            self.filesavedialog = None
+        self.newDocument()
+
+    def exportStaplesCallback(self, selected: Union[str, list, tuple]):
+        '''Export all staple sequences to selected CSV file.
+
+        Args:
+            selected (Tuple, List or str): if a List or Tuple, the filename should
+            be the first element
+        '''
+        if isinstance(selected, (list, tuple)):
+            fname = selected[0]
+        else:
+            fname = selected
+        # Return if fname is '', None, or a directory path
+        if not fname or fname is None or os.path.isdir(fname):
+            return False
+        if not fname.lower().endswith(".txt"):
+            fname += ".txt"
+        if self.saveStaplesDialog is not None:
+            self.saveStaplesDialog.filesSelected.disconnect(self.exportStaplesCallback)
+            # manual garbage collection to prevent hang (in osx)
+            del self.saveStaplesDialog
+            self.saveStaplesDialog = None
+        # write the file
+        ap = self._document.activePart()
+        if ap is not None:
+            output = ap.getSequences()
+            with open(fname, 'w') as f:
+                f.write(output)
+    # end def
+
+    def newClickedCallback(self):
+        '''Gets called on completion of filesavedialog after newClicked's
+        promptSaveDialog. Removes the dialog if necessary, but it was probably
+        already removed by saveFileDialogCallback.
+        '''
+
+        if self.filesavedialog is not None:
+            self.filesavedialog.finished.disconnect(self.newClickedCallback)
+            del self.filesavedialog  # prevents hang (?)
+            self.filesavedialog = None
+        self.newDocument()
+
+    def openAfterMaybeSaveCallback(self, selected: Union[str, list, tuple]):
+        '''Receives file selection info from the dialog created by
+        openAfterMaybeSave, following user input.
+
+        Extracts the file name and passes it to the decode method, which
+        returns a new document doc, which is then set as the open document
+        by newDocument. Calls finalizeImport and disconnects dialog signaling.
+        '''
+        if isinstance(selected, (list, tuple)):
+            fname = selected[0]
+        else:
+            fname = selected
+        if fname is None or fname == '' or os.path.isdir(fname):
+            return False
+        if not os.path.exists(fname):
+            return False
+        self._writeFileOpenPath(os.path.dirname(fname))
+
+        win = self.win
+        win.path_graphics_view.setViewportUpdateOn(False)
+        win.slice_graphics_view.setViewportUpdateOn(False)
+        win.grid_graphics_view.setViewportUpdateOn(False)
+
+        if ONLY_ONE:
+            self.newDocument(fname=fname)
+
+        self._document.readFile(fname)
+
+        win.path_graphics_view.setViewportUpdateOn(True)
+        win.slice_graphics_view.setViewportUpdateOn(True)
+        win.grid_graphics_view.setViewportUpdateOn(True)
+
+        win.path_graphics_view.update()
+        win.slice_graphics_view.update()
+        win.grid_graphics_view.update()
+
+        if hasattr(self, "filesavedialog"):  # user did save
+            if self.fileopendialog is not None:
+                self.fileopendialog.filesSelected.disconnect(self.openAfterMaybeSaveCallback)
+            # manual garbage collection to prevent hang (in osx)
+            del self.fileopendialog
+            self.fileopendialog = None
+        self.setFileName(fname)
+
+    def saveFileDialogCallback(self, selected: Union[str, list, tuple]):
+        '''If the user chose to save, write to that file.'''
+        if isinstance(selected, (list, tuple)):
+            fname = selected[0]
+        else:
+            fname = selected
+        if fname is None or os.path.isdir(fname):
+            return False
+        if not fname.lower().endswith(".json"):
+            fname += ".json"
+        if self.filesavedialog is not None:
+            self.filesavedialog.filesSelected.disconnect(self.saveFileDialogCallback)
+            del self.filesavedialog  # prevents hang
+            self.filesavedialog = None
+        self.writeDocumentToFile(fname)
+        self._writeFileOpenPath(os.path.dirname(fname))
+
+        if self.exit_when_done:
+            the_app = app()
+            self.destroyDC()
+            if the_app.document_controllers:
+                the_app.destroyApp()
+
+    ### EVENT HANDLERS ###
+    def windowCloseEventHandler(self, event: QCloseEvent = None):
+        '''Intercept close events when user attempts to close the window.'''
+        win = self.win
+        dialog_result = self.promptSaveDialog(exit_when_done=True)
+        if dialog_result == SAVE_DIALOG_OPTIONS['DISCARD']:
+            if event is not None:
+                event.accept()
+            the_app = app()
+            self.destroyDC()
+            if the_app.document_controllers:
+                the_app.destroyApp()
+        elif event is not None:
+            event.ignore()
+    # end def
+
+    ### FILE INPUT ##
+    def documentTitle(self) -> str:
+        fname = os.path.basename(str(self.fileName()))
+        if not self.undoStack().isClean():
+            fname += '[*]'
+        return fname
+
+    def fileName(self) -> str:
+        return self._document.fileName()
+
+    def setFileName(self, proposed_fname: str) -> bool:
+        if self.fileName() == proposed_fname:
+            return True
+        self._document.setFileName(proposed_fname)
+        self._has_no_associated_file = False
+        self.win.setWindowTitle(self.documentTitle())
+        return True
+
+    def openAfterMaybeSave(self):
+        '''This is the method that initiates file opening. It is called by
+        actionOpenSlot to spawn a QFileDialog and connect it to a callback
+        method.
+        '''
+        path = self._file_open_path
+        if util.isWindows():  # required for native looking file window#"/",
+            fname = QFileDialog.getOpenFileName(None,
+                                                "Open Document", path,
+                                                "cadnano1 / cadnano2 Files (*.nno *.json *.c25)")
+            self.filesavedialog = None
+            self.openAfterMaybeSaveCallback(fname)
+        else:  # access through non-blocking callback
+            fdialog = QFileDialog(self.win,
+                                  "Open Document",
+                                  path,
+                                  "cadnano1 / cadnano2 Files (*.nno *.json *.c25)")
+            fdialog.setAcceptMode(QFileDialog.AcceptOpen)
+            fdialog.setWindowFlags(Qt.Sheet)
+            fdialog.setWindowModality(Qt.WindowModal)
+            self.fileopendialog = fdialog
+            self.fileopendialog.filesSelected.connect(self.openAfterMaybeSaveCallback)
+            fdialog.open()
+    # end def
+
+    ### FILE OUTPUT ###
+    def promptSaveDialog(self, exit_when_done: bool = False):
+        '''Save on quit, check if document changes have occurred.
+
+        Returns:
+            SAVE_DIALOG_OPTIONS['CANCEL'] or
+            SAVE_DIALOG_OPTIONS['DISCARD'] or
+            SAVE_DIALOG_OPTIONS['SAVE']
+        '''
+        if app().dontAskAndJustDiscardUnsavedChanges:
+            return SAVE_DIALOG_OPTIONS['DISCARD']
+        if not self.undoStack().isClean():    # document dirty?
+            savebox = QMessageBox(QMessageBox.Warning, "Application",
+                                  "The document has been modified.\nDo you want to save your changes?",
+                                  QMessageBox.Save | QMessageBox.Discard | QMessageBox.Cancel,
+                                  self.win,
+                                  Qt.Dialog | Qt.MSWindowsFixedSizeDialogHint | Qt.Sheet)
+            savebox.setWindowModality(Qt.WindowModal)
+            save = savebox.button(QMessageBox.Save)
+            discard = savebox.button(QMessageBox.Discard)
+            cancel = savebox.button(QMessageBox.Cancel)
+            savebox.setDefaultButton(save)
+            savebox.setEscapeButton(cancel)
+            save.setShortcut("Ctrl+S")
+            discard.setShortcut(QKeySequence("D,Ctrl+D"))
+            cancel.setShortcut(QKeySequence("C,Ctrl+C,.,Ctrl+."))
+            ret = savebox.exec_()
+            del savebox  # manual garbage collection to prevent hang (in osx)
+            if ret == QMessageBox.Save:
+                self.exit_when_done = exit_when_done
+                self.actionSaveAsSlot()
+                return SAVE_DIALOG_OPTIONS['SAVE']
+            elif ret == QMessageBox.Cancel:
+                return SAVE_DIALOG_OPTIONS['CANCEL']
+        return SAVE_DIALOG_OPTIONS['DISCARD']
+
+    def writeDocumentToFile(self, filename: str = None):
+        if filename is None or filename == '':
+            if self._has_no_associated_file:
+                return False
+            filename = self.fileName()
+        try:
+            self._document.writeToFile(filename)
+        except Exception:
+            flags = Qt.Dialog | Qt.MSWindowsFixedSizeDialogHint | Qt.Sheet
+            errorbox = QMessageBox(QMessageBox.Critical,
+                                   "cadnano",
+                                   "Could not write to '%s'." % filename,
+                                   QMessageBox.Ok,
+                                   self.win,
+                                   flags)
+            errorbox.setWindowModality(Qt.WindowModal)
+            errorbox.open()
+            raise
+            return False
+        self.undoStack().setClean()
+        self.setFileName(filename)
+        return True
+
+    def actionCadnanoWebsiteSlot(self):
+        import webbrowser
+        webbrowser.open("http://cadnano.org/")
+
+    def actionFeedbackSlot(self):
+        import webbrowser
+        webbrowser.open("http://cadnano.org/feedback")
 
 
 # end class
